+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v1.0.11
+* Feature: Added copy prices text to clipboard on the Trading Calculator.
+* Feature: Added copy total text to clipboard on the Compound Interest Calculator.
+
 ## v1.0.10
 * Feature: Signing with Google Account.
 * Feature: Dashboard available to see interest data.